Added default data when initially installing app.
Reversed the display order of journal entries so the newest is shown first in the journal view.

Refactored journal ID from Int to Long throughout whole project.

+++ app/src/main/java/com/delly/journeyjournal/NavigationManager.kt
@@ -15,10 +15,10 @@ import kotlinx.serialization.Serializable
 object Home
 
 @Serializable
-data class CreateEditJourney(val journalId: Int? = null)
+data class CreateEditJourney(val journalId: Long? = null)
 
 @Serializable
-data class SelectedJourney(val id: Int)
+data class SelectedJourney(val id: Long)
 
 /**
  * Navigation graph for the app.
