+++ app/src/main/java/com/delly/journeyjournal/journalUi/JournalEntriesNav.kt
@@ -22,7 +22,7 @@ object CreateJourneyEntryDestination
 @Composable
 fun JournalEntriesNav(
     repository: JournalRepository,
-    journalId: Int
+    journalId: Long
 ) {
     val navController = rememberNavController()
