+++ app/src/main/java/com/delly/journeyjournal/viewmodels/JournalEntriesViewModel.kt
@@ -17,7 +17,7 @@ import kotlinx.coroutines.flow.stateIn
  */
 class JournalEntriesViewModel(
     private val repository: JournalRepository,
-    private val journalId: Int,
+    private val journalId: Long,
 ) : ViewModel() {
 
     /**
